refactor(login): move login to a separate component and rearrange
